Convert all ECC implementations to the new ASN code and share it.
